requirements.txt
torch~=2.6.0.dev20240930+cu124
torchaudio~=2.5.0.dev20240930+cu124
torchvision~=0.20.0.dev20240930+cu124
transformers~=4.45.1
pillow~=10.4.0
tokenizers~=0.20.0
qwen_vl_utils~=0.0.8
safetensors~=0.4.5